requirements.txt
python-telegram-bot==20.7
openai==1.12.0
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
jinja2==3.1.3 